Attacking bug resolved and several other things. Details in description.
Since each tower was constructing an enemies list only at the beginning of the game (in Start()) there was only one enemy in each towers enemies list and they only attacked that enemy. Instead of this stupid way, made wavespawner keep track of active enemies it spawns and expose this list of enemies to towers.
Implemented observer pattern to let the wavespawner and towers know that an enemy died and act accordingly.
Made wave spawner cache the Enemy components of spawned enemies instead of using GetComponent<>().