Another tower attacking bug fixed & target acquiring & attack revised.
Bug: Towers were attacking the first enemy they acquired until it died simply because there were no checks to look for distance between the tower and the enemy.

Since the number of towers and enemies can be high, eliminating target acquiring and attacking checks from the update method might prove useful.

Target acquiring and attacking is no longer handled by the Update method. Instead, in the start method, an InvokeRepeating is issued for AcquireTarget method. It checks for new targets every 0.1 seconds and when it finds one, it cancels it InvokeRepeating and issues a new one for AttackWhenReady. AttackWhenReady checks for cooldown status and attacks accordingly each 0.1 seconds. When the enemy dies or is out of range, AttackWhenReady does the same and cancels its InvokeRepeating call, issuing a new one for AcquireTarget and effectively creating a cycle.

Cooldown is now handled via Cooldown coroutine now. It decrements a timer ten times and resets the cooldown status.